requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Twilio==0.0.6
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
twilio==6.51.1
urllib3==1.26.3
Werkzeug==1.0.1
